+++ apps/native-component-list/src/screens/UI/TextInputScreen.ios.tsx
@@ -126,9 +126,9 @@ export default function TextInputScreen() {
           <Picker
             label="Submit label"
             modifiers={[pickerStyle('menu')]}
-            onSelectionChange={({ nativeEvent: { selection } }) => {
-              setSelectedSubmitLabel(selection as (typeof submitLabelOptions)[number]);
-            }}>
+            onSelectionChange={(selection) =>
+              setSelectedSubmitLabel(selection as (typeof submitLabelOptions)[number])
+            }>
             {submitLabelOptions.map((option) => (
               <Text key={option} modifiers={[tag(option)]}>
                 {option}